Implement packet codec system to remove expensive protocol state switches
With the way this lib used to previously work, any time the protocol state changed, all the packet maps got cleared and had new packets inserted into them depending on the protocol state. This was unnecessarily expensive and although called rarely, it was not ideal.

The packet codecs added are static, immutable and can be reused across sessions. The codec system also opens up the door for multi-version capability in the future.