expose the shutdown method of the event loop group
To exit normally all non-daemon threads of the java process are supposed
to have stopped. TcpClientSession creates an static class member
EventLoopGroup which in turn instantiates worker threads. These need to
be stopped as a part of shutting down the server.

This patch exposes the shutdown of the EventLoopGroup as a static
function intended to be called by some central function as a part of
server shutdwon.

Nullify the EventLoopGroup in preparation for starting new users
of TcpClientSession.